test(filters): add snapshot tests for filtering of all doc samples
