+++ pasting/dropboxpaper-ie11-unsupported-strippastedstyles-win81.json
@@ -1,8 +1,4 @@
 {
-    "metadata": {
-        "error": "Unsupported browser",
-        "stripPastedStyles": true
-    },
     "entityMap": {},
     "blocks": [
         {
