+++ pasting/dropboxpaper-ie11-unsupported-win81.json
@@ -1,7 +1,4 @@
 {
-    "metadata": {
-        "error": "Unsupported browser"
-    },
     "entityMap": {},
     "blocks": [
         {
