+++ pasting/word2010-ie11-noequation-strippastedstyles-win81.json
@@ -1,8 +1,4 @@
 {
-    "metadata": {
-        "error": "Removed equation content (browser crash)",
-        "stripPastedStyles": true
-    },
     "entityMap": {},
     "blocks": [
         {
