--- pasting/wordonline-ie11-win81.json
@@ -1,5 +1,0 @@
-{
-    "metadata": {
-        "error": "Unsupported browser"
-    }
-}
