+++ src/tests/fixtures.js
@@ -1,0 +1,47 @@
+export const microsoftWord = {
+  "word2010-chrome62-win81": require("../../pasting/word2010-chrome62-win81.json"),
+  "word2010-firefox57-win81": require("../../pasting/word2010-firefox57-win81.json"),
+  "word2010-ie11-noequation-strippastedstyles-win81": require("../../pasting/word2010-ie11-noequation-strippastedstyles-win81.json"),
+  "word2010-ie11-noequation-win81": require("../../pasting/word2010-ie11-noequation-win81.json"),
+  "word-safari-ios11": require("../../pasting/word-safari-ios11.json"),
+}
+
+export const googleDocs = {
+  "googledocs-chrome62-macos1013": require("../../pasting/googledocs-chrome62-macos1013.json"),
+  "googledocs-chrome62-win81": require("../../pasting/googledocs-chrome62-win81.json"),
+  "googledocs-edge16-win10": require("../../pasting/googledocs-edge16-win10.json"),
+  "googledocs-firefox57-macos1013": require("../../pasting/googledocs-firefox57-macos1013.json"),
+  "googledocs-firefox57-win81": require("../../pasting/googledocs-firefox57-win81.json"),
+  "googledocs-ie11-strippastedstyles-win81": require("../../pasting/googledocs-ie11-strippastedstyles-win81.json"),
+  "googledocs-ie11-win81": require("../../pasting/googledocs-ie11-win81.json"),
+  "googledocs-safari-ios11": require("../../pasting/googledocs-safari-ios11.json"),
+  "googledocs-safari11-macos1013": require("../../pasting/googledocs-safari11-macos1013.json"),
+}
+
+export const dropboxPaper = {
+  "dropboxpaper-chrome62-macos1013": require("../../pasting/dropboxpaper-chrome62-macos1013.json"),
+  "dropboxpaper-chrome62-win81": require("../../pasting/dropboxpaper-chrome62-win81.json"),
+  "dropboxpaper-edge16-win10": require("../../pasting/dropboxpaper-edge16-win10.json"),
+  "dropboxpaper-firefox57-macos1013": require("../../pasting/dropboxpaper-firefox57-macos1013.json"),
+  "dropboxpaper-firefox57-win81": require("../../pasting/dropboxpaper-firefox57-win81.json"),
+  "dropboxpaper-ie11-unsupported-strippastedstyles-win81": require("../../pasting/dropboxpaper-ie11-unsupported-strippastedstyles-win81.json"),
+  "dropboxpaper-ie11-unsupported-win81": require("../../pasting/dropboxpaper-ie11-unsupported-win81.json"),
+  "dropboxpaper-safari-ios11": require("../../pasting/dropboxpaper-safari-ios11.json"),
+  "dropboxpaper-safari11-macos1013": require("../../pasting/dropboxpaper-safari11-macos1013.json"),
+}
+
+export const applePages = {
+  "applepages-chrome62-macos1013": require("../../pasting/applepages-chrome62-macos1013.json"),
+  "applepages-firefox57-macos1013": require("../../pasting/applepages-firefox57-macos1013.json"),
+  "applepages-safari-ios11": require("../../pasting/applepages-safari-ios11.json"),
+  "applepages-safari11-macos1013": require("../../pasting/applepages-safari11-macos1013.json"),
+}
+
+export const wordOnline = {
+  "wordonline-chrome62-macos1013": require("../../pasting/wordonline-chrome62-macos1013.json"),
+  "wordonline-chrome62-win81": require("../../pasting/wordonline-chrome62-win81.json"),
+  "wordonline-edge16-win10": require("../../pasting/wordonline-edge16-win10.json"),
+  "wordonline-firefox57-macos1013": require("../../pasting/wordonline-firefox57-macos1013.json"),
+  "wordonline-firefox57-win81": require("../../pasting/wordonline-firefox57-win81.json"),
+  "wordonline-safari11-macos1013": require("../../pasting/wordonline-safari11-macos1013.json"),
+}
