+++ src/tests/pages.test.js
@@ -1,0 +1,8 @@
+import { testFilteringDiff } from "./utils"
+import { applePages } from "./fixtures"
+
+describe("Apple Pages", () => {
+  Object.keys(applePages).forEach((key) => {
+    it(key, () => testFilteringDiff(applePages[key]))
+  })
+})
