+++ src/tests/word.test.js
@@ -1,0 +1,8 @@
+import { testFilteringDiff } from "./utils"
+import { microsoftWord } from "./fixtures"
+
+describe("Microsoft Word", () => {
+  Object.keys(microsoftWord).forEach((key) => {
+    it(key, () => testFilteringDiff(microsoftWord[key]))
+  })
+})
